1. fix ALS bug 2. add unit test for converting buffer to CSRNumericTable
